requirements.txt
openai
elevenlabs
moviepy
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client